Data Validation

+++ backend/StockPortfolio.API/Controllers/CommentController.cs
@@ -25,15 +25,21 @@ namespace StockPortfolio.API.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments = await _commentRepository.GetAllAsync();
             var commentDtos = comments.Select(c => c.ToCommentDto());
 
             return Ok(commentDtos);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById(int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             Comment comment = await _commentRepository.GetByIdAsync(id);
 
             if (comment == null)
@@ -44,12 +50,15 @@ namespace StockPortfolio.API.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create(
             [FromRoute] int stockId,
             CreateCommentDto commentDto
         )
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             if (!await this._stockRepository.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist");
@@ -65,9 +74,31 @@ namespace StockPortfolio.API.Controllers
             );
         }
 
-        [HttpDelete("{id}")]
+        [HttpPut("{id:int}")]
+        public async Task<IActionResult> Update(
+            [FromRoute] int id,
+            [FromBody] UpdateCommentDto commentDto
+        )
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            Comment? comment = await _commentRepository.UpdateAsync(id, commentDto);
+
+            if (comment == null)
+            {
+                return NotFound();
+            }
+
+            return NoContent();
+        }
+
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var commentModel = await _commentRepository.DeleteAsync(id);
 
             if (commentModel == null)
